[skip ci] add version in separate script
